Berkeley DB 12cR1 (12.1.6.2.38)
`lang/sql/sqlite/tool` is removed.
Cannot follow the [SQLite Amalgamation guide](https://www.sqlite.org/amalgamation.html) now.

<https://download.oracle.com/otndocs/products/berkeleydb/html/changelog_6_2.html>

<https://download.oracle.com/otn/berkeley-db/db-6.2.38.tar.gz>

+++ examples/cxx/EnvExample.cpp
@@ -88,7 +88,7 @@ db_setup(const char *home, const char *data_dir, ostream& err_stream)
 	// Create an environment object and initialize it for error
 	// reporting.
 	//
-	DbEnv *dbenv = new DbEnv(0);
+	DbEnv *dbenv = new DbEnv((u_int32_t)0);
 	dbenv->set_error_stream(&err_stream);
 	dbenv->set_errpfx(progname);
 
@@ -148,7 +148,7 @@ void
 db_teardown(const char *home, const char *data_dir, ostream& err_stream)
 {
 	// Remove the shared database regions.
-	DbEnv *dbenv = new DbEnv(0);
+	DbEnv *dbenv = new DbEnv((u_int32_t)0);
 
 	dbenv->set_error_stream(&err_stream);
 	dbenv->set_errpfx(progname);
